aReclambord_server/requirements.txt
asgiref==3.2.7
bson==0.5.8
click==7.1.2
dataclasses==0.6
Django==2.2.13
djangorestframework==3.11.0
djongo==1.3.2
dnspython==1.16.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.2.0
pymongo==3.10.1
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2019.3
six==1.15.0
sqlparse==0.2.4
Werkzeug==1.0.1
